Adding Juno.Runtime.Protocol.Types separating Runtime.Types and Runtime.Protocol.Types into the types that are shared between the API and the juno/raft procotol only types.
