+++ src/Juno/Runtime/Protocol/Types.hs
@@ -1,0 +1,711 @@
+{-# LANGUAGE DeriveTraversable #-}
+{-# LANGUAGE TypeFamilies #-}
+{-# OPTIONS_GHC -fno-warn-orphans #-}
+{-# LANGUAGE BangPatterns #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE StandaloneDeriving #-}
+{-# LANGUAGE FunctionalDependencies #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE GeneralizedNewtypeDeriving #-}
+{-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE Rank2Types #-}
+{-# LANGUAGE RecordWildCards #-}
+{-# LANGUAGE TemplateHaskell #-}
+
+-----------------------------------------------------------------------------
+-- |
+-- Module      :  Juno.Runtime.Protocol.Types
+--
+-- Holds the core Juno/Raft Types used to implement BFT Raft
+-- the types here are internal to the protocol nodes, but for now they share
+-- some types Runtime/Types.hs with Api/Types.hs.
+
+module Juno.Runtime.Protocol.Types
+  ( Raft
+  , RaftSpec(..)
+  , readLogEntry, writeLogEntry, readTermNumber, writeTermNumber
+  , readVotedFor, writeVotedFor, applyLogEntry, sendMessage
+  , sendMessages, getMessage, getMessages, getNewCommands, getNewEvidence
+  , debugPrint, publishMetric, getTimestamp, random
+  , enqueue, enqueueMultiple, dequeue, enqueueLater, killEnqueued
+  , NodeID(..)
+  , CommandEntry(..)
+  , CommandResult(..)
+  , CommandStatus(..)
+  , Term(..), startTerm
+  , LogIndex(..), startIndex
+  , RequestId(..), startRequestId, toRequestId
+  , Config(..), otherNodes, nodeId, electionTimeoutRange, heartbeatTimeout
+  , enableDebug, publicKeys, clientPublicKeys, myPrivateKey, clientTimeoutLimit
+  , myPublicKey, batchTimeDelta
+  , Role(..)
+  , Metric(..)
+  , RaftEnv(..), cfg, clusterSize, quorumSize, rs
+  , LogEntry(..), leTerm, leCommand, leHash
+  , Log(..), lEntries
+  , RaftState(..), role, term, votedFor, lazyVote, currentLeader, ignoreLeader
+  , logEntries, commitIndex, commitProof, lastApplied, timerThread, replayMap
+  , cYesVotes, cPotentialVotes, lNextIndex, lMatchIndex, lConvinced
+  , lastCommitTime, numTimeouts, pendingRequests, currentRequestId
+  , timeSinceLastAER, lLastBatchUpdate
+  , initialRaftState
+  -- * RPC
+  , AppendEntries(..)
+  , AppendEntriesResponse(..), AERWire(..), AlotOfAERs(..)
+  , RequestVote(..)
+  , RequestVoteResponse(..)
+  , Command(..)
+  , CommandResponse(..)
+  , CommandBatch(..)
+  , Revolution(..)
+  , RPC(..)
+  , Event(..)
+  , KeySet(..), WireFormat(..)
+  , signedRPCtoRPC, rpcToSignedRPC
+  , SignedRPC(..)
+  -- for simplicity, re-export some core types that we need all over the place
+  , PublicKey, PrivateKey, Signature(..), sign, valid, importPublic, importPrivate
+  -- for testing & benchmarks
+  , LEWire(..), encodeLEWire, decodeLEWire, decodeRVRWire
+  , verifySignedRPC, CMDWire(..)
+  ) where
+
+import Control.Monad (mzero)
+import Control.Parallel.Strategies
+import Control.Concurrent (ThreadId)
+import Control.Lens hiding (Index, (|>))
+import qualified Control.Lens as Lens
+import Control.Monad.RWS (RWST)
+--import Crypto.Ed25519.Pure ( PublicKey, PrivateKey, Signature(..), sign, valid
+--                           , importPublic, importPrivate, exportPublic, exportPrivate)
+import Data.Sequence (Seq, (|>))
+import qualified Data.Sequence as Seq
+import Data.Map (Map)
+import qualified Data.Map as Map
+import Data.Set (Set)
+import qualified Data.Set as Set
+import Data.ByteString (ByteString)
+import Data.Serialize (Serialize)
+import qualified Data.Serialize as S
+import Data.Foldable
+import Text.Read (readMaybe)
+import qualified Data.Text as Text
+import Data.Thyme.Clock
+import Data.Thyme.Time.Core ()
+
+import Data.Aeson (genericParseJSON,genericToJSON,parseJSON,toJSON,ToJSON,FromJSON,Value(..))
+import Data.Aeson.Types (defaultOptions,Options(..))
+
+import GHC.Generics hiding (from)
+
+import System.Random (Random)
+import Juno.Runtime.Types
+
+newtype Term = Term Int
+  deriving (Show, Read, Eq, Enum, Num, Ord, Generic, Serialize)
+
+startTerm :: Term
+startTerm = Term (-1)
+
+newtype LogIndex = LogIndex Int
+  deriving (Show, Read, Eq, Ord, Enum, Num, Real, Integral, Generic, Serialize)
+
+startIndex :: LogIndex
+startIndex = LogIndex (-1)
+
+data Config = Config
+  { _otherNodes           :: !(Set NodeID)
+  , _nodeId               :: !NodeID
+  , _publicKeys           :: !(Map NodeID PublicKey)
+  , _clientPublicKeys     :: !(Map NodeID PublicKey)
+  , _myPrivateKey         :: !PrivateKey
+  , _myPublicKey          :: !PublicKey
+  , _electionTimeoutRange :: !(Int,Int)
+  , _heartbeatTimeout     :: !Int
+  , _batchTimeDelta       :: !NominalDiffTime
+  , _enableDebug          :: !Bool
+  , _clientTimeoutLimit   :: !Int
+  }
+  deriving (Show, Generic)
+makeLenses ''Config
+instance ToJSON NominalDiffTime where
+  toJSON = toJSON . show . toSeconds'
+instance FromJSON NominalDiffTime where
+  parseJSON (String s) = case readMaybe $ Text.unpack s of
+    Just s' -> return $ fromSeconds' s'
+    Nothing -> mzero
+  parseJSON _ = mzero
+instance ToJSON Config where
+  toJSON = genericToJSON defaultOptions { fieldLabelModifier = drop 1 }
+instance FromJSON Config where
+  parseJSON = genericParseJSON defaultOptions { fieldLabelModifier = drop 1 }
+
+data KeySet = KeySet
+  { _ksCluster :: !(Map NodeID PublicKey)
+  , _ksClient  :: !(Map NodeID PublicKey)
+  } deriving (Show)
+
+-- | Type that is serialized and sent over the wire
+data SignedRPC = SignedRPC
+  { _sigDigest :: !Digest
+  , _sigBody   :: !ByteString
+  } deriving (Show, Eq, Generic)
+instance Serialize SignedRPC
+
+-- | Based on the MsgType in the SignedRPC's Digest, we know which set of keys are needed to validate the message
+verifySignedRPC :: KeySet -> SignedRPC -> Either String ()
+verifySignedRPC !KeySet{..} s@(SignedRPC !Digest{..} !bdy)
+  | _digType == CMD || _digType == REV || _digType == CMDB =
+      case Map.lookup _digNodeId _ksClient of
+        Nothing -> Left $! "PubKey not found for NodeID: " ++ show _digNodeId
+        Just !key
+          | key /= _digPubkey -> Left $! "Public key in storage doesn't match digest's key for msg: " ++ show s
+          | otherwise -> if not $ valid bdy key _digSig
+                         then Left $! "Unable to verify SignedRPC sig: " ++ show s
+                         else Right ()
+  | otherwise =
+      case Map.lookup _digNodeId _ksCluster of
+        Nothing -> Left $! "PubKey not found for NodeID: " ++ show _digNodeId
+        Just !key
+          | key /= _digPubkey -> Left $! "Public key in storage doesn't match digest's key for msg: " ++ show s
+          | otherwise -> if not $ valid bdy key _digSig
+                         then Left $! "Unable to verify SignedRPC sig: " ++ show s
+                         else Right ()
+{-# INLINE verifySignedRPC #-}
+
+class WireFormat a where
+  toWire   :: NodeID -> PublicKey -> PrivateKey -> a -> SignedRPC
+  fromWire :: Maybe ReceivedAt -> KeySet -> SignedRPC -> Either String a
+
+data CMDWire = CMDWire !(CommandEntry, NodeID, RequestId)
+  deriving (Show, Generic)
+instance Serialize CMDWire
+
+instance WireFormat Command where
+  toWire nid pubKey privKey Command{..} = case _cmdProvenance of
+    NewMsg -> let bdy = S.encode $ CMDWire (_cmdEntry, _cmdClientId, _cmdRequestId)
+                  sig = sign bdy privKey pubKey
+                  dig = Digest nid sig pubKey CMD
+              in SignedRPC dig bdy
+    ReceivedMsg{..} -> SignedRPC _pDig _pOrig
+  fromWire !ts !ks s@(SignedRPC !dig !bdy) =
+    case verifySignedRPC ks s of
+      Left !err -> Left err
+      Right () -> if _digType dig /= CMD
+        then error $ "Invariant Failure: attempting to decode " ++ show (_digType dig) ++ " with CMDWire instance"
+        else case S.decode bdy of
+            Left !err -> Left $! "Failure to decode CMDWire: " ++ err
+            Right (CMDWire !(ce,nid,rid)) -> Right $! Command ce nid rid $ ReceivedMsg dig bdy ts
+  {-# INLINE toWire #-}
+  {-# INLINE fromWire #-}
+
+instance WireFormat CommandBatch where
+  toWire nid pubKey privKey CommandBatch{..} = case _cmdbProvenance of
+    NewMsg -> let bdy = S.encode ((toWire nid pubKey privKey <$> _cmdbBatch) `using` parList rseq)
+                  sig = sign bdy privKey pubKey
+                  dig = Digest nid sig pubKey CMDB
+              in SignedRPC dig bdy
+    ReceivedMsg{..} -> SignedRPC _pDig _pOrig
+  fromWire !ts !ks s@(SignedRPC dig bdy) = case verifySignedRPC ks s of
+    Left !err -> Left err
+    Right () -> if _digType dig /= CMDB
+      then error $! "Invariant Failure: attempting to decode " ++ show (_digType dig) ++ " with CMDBWire instance"
+      else case S.decode bdy of
+        Left !err -> Left $ "Failure to decode CMDBWire: " ++ err
+        Right !cmdb' -> gatherValidCmdbs (ReceivedMsg dig bdy ts) ((fromWire ts ks <$> cmdb') `using` parList rseq)
+  {-# INLINE toWire #-}
+  {-# INLINE fromWire #-}
+
+gatherValidCmdbs :: Provenance -> [Either String Command] -> Either String CommandBatch
+gatherValidCmdbs prov ec = (`CommandBatch` prov) <$> sequence ec
+{-# INLINE gatherValidCmdbs #-}
+
+data CMDRWire = CMDRWire (CommandResult, NodeID, NodeID, RequestId)
+  deriving (Show, Generic)
+instance Serialize CMDRWire
+
+instance WireFormat CommandResponse where
+  toWire nid pubKey privKey CommandResponse{..} = case _cmdrProvenance of
+    NewMsg -> let bdy = S.encode $ CMDRWire (_cmdrResult,_cmdrLeaderId,_cmdrNodeId,_cmdrRequestId)
+                  sig = sign bdy privKey pubKey
+                  dig = Digest nid sig pubKey CMDR
+              in SignedRPC dig bdy
+    ReceivedMsg{..} -> SignedRPC _pDig _pOrig
+  fromWire ts !ks s@(SignedRPC !dig !bdy) = case verifySignedRPC ks s of
+    Left !err -> Left err
+    Right () -> if _digType dig /= CMDR
+      then error $ "Invariant Failure: attempting to decode " ++ show (_digType dig) ++ " with CMDRWire instance"
+      else case S.decode bdy of
+        Left !err -> Left $! "Failure to decode CMDRWire: " ++ err
+        Right (CMDRWire !(r,lid,nid,rid)) -> Right $! CommandResponse r lid nid rid $ ReceivedMsg dig bdy ts
+  {-# INLINE toWire #-}
+  {-# INLINE fromWire #-}
+
+data LogEntry = LogEntry
+  { _leTerm    :: !Term
+  , _leCommand :: !Command
+  , _leHash    :: !ByteString
+  }
+  deriving (Show, Eq, Generic)
+makeLenses ''LogEntry
+
+newtype Log a = Log { _lEntries :: Seq a }
+    deriving (Eq,Show,Ord,Generic,Monoid,Functor,Foldable,Traversable,Applicative,Monad,NFData)
+makeLenses ''Log
+instance (t ~ Log a) => Rewrapped (Log a) t
+instance Wrapped (Log a) where
+    type Unwrapped (Log a) = Seq a
+    _Wrapped' = iso _lEntries Log
+instance Cons (Log a) (Log a) a a where
+    _Cons = _Wrapped . _Cons . mapping _Unwrapped
+instance Snoc (Log a) (Log a) a a where
+    _Snoc = _Wrapped . _Snoc . firsting _Unwrapped
+type instance IxValue (Log a) = a
+type instance Lens.Index (Log a) = LogIndex
+instance Ixed (Log a) where ix i = lEntries.ix (fromIntegral i)
+
+data LEWire = LEWire (Term, SignedRPC, ByteString)
+  deriving (Show, Generic)
+instance Serialize LEWire
+
+decodeLEWire' :: Maybe ReceivedAt -> KeySet -> LEWire -> Either String LogEntry
+decodeLEWire' !ts !ks (LEWire !(t,cmd,hsh)) = case fromWire ts ks cmd of
+      Left !err -> Left $!err
+      Right !cmd' -> Right $! LogEntry t cmd' hsh
+{-# INLINE decodeLEWire' #-}
+
+-- TODO: check if `toSeqLogEntry ele = Seq.fromList <$> sequence ele` is fusable?
+toSeqLogEntry :: [Either String LogEntry] -> Either String (Seq LogEntry)
+toSeqLogEntry !ele = go ele mempty
+  where
+    go [] s = Right $! s
+    go (Right le:les) s = go les (s |> le)
+    go (Left err:_) _ = Left $! err
+{-# INLINE toSeqLogEntry #-}
+
+decodeLEWire :: Maybe ReceivedAt -> KeySet -> [LEWire] -> Either String (Seq LogEntry)
+decodeLEWire !ts !ks !les = go les Seq.empty
+  where
+    go [] s = Right $! s
+    go (LEWire !(t,cmd,hsh):ls) v = case fromWire ts ks cmd of
+      Left err -> Left $! err
+      Right cmd' -> go ls (v |> LogEntry t cmd' hsh)
+{-# INLINE decodeLEWire #-}
+
+encodeLEWire :: NodeID -> PublicKey -> PrivateKey -> Seq LogEntry -> [LEWire]
+encodeLEWire nid pubKey privKey les =
+  (\LogEntry{..} -> LEWire (_leTerm, toWire nid pubKey privKey _leCommand, _leHash)) <$> toList les
+{-# INLINE encodeLEWire #-}
+
+data AppendEntries = AppendEntries
+  { _aeTerm        :: !Term
+  , _leaderId      :: !NodeID
+  , _prevLogIndex  :: !LogIndex
+  , _prevLogTerm   :: !Term
+  , _aeEntries     :: !(Seq LogEntry)
+  , _aeQuorumVotes :: !(Set RequestVoteResponse)
+  , _aeProvenance  :: !Provenance
+  }
+  deriving (Show, Eq, Generic)
+
+data AEWire = AEWire (Term,NodeID,LogIndex,Term,[LEWire],[SignedRPC])
+  deriving (Show, Generic)
+instance Serialize AEWire
+
+instance WireFormat AppendEntries where
+  toWire nid pubKey privKey AppendEntries{..} = case _aeProvenance of
+    NewMsg -> let bdy = S.encode $ AEWire (_aeTerm
+                                          ,_leaderId
+                                          ,_prevLogIndex
+                                          ,_prevLogTerm
+                                          ,encodeLEWire nid pubKey privKey _aeEntries
+                                          ,toWire nid pubKey privKey <$> toList _aeQuorumVotes)
+                  sig = sign bdy privKey pubKey
+                  dig = Digest nid sig pubKey AE
+              in SignedRPC dig bdy
+    ReceivedMsg{..} -> SignedRPC _pDig _pOrig
+  fromWire !ts !ks s@(SignedRPC !dig !bdy) = case verifySignedRPC ks s of
+    Left !err -> Left err
+    Right () -> if _digType dig /= AE
+      then error $ "Invariant Failure: attempting to decode " ++ show (_digType dig) ++ " with AEWire instance"
+      else case S.decode bdy of
+        Left err -> Left $! "Failure to decode AEWire: " ++ err
+        Right (AEWire (t,lid,pli,pt,les,vts)) -> runEval $ do
+          eLes <- rpar (toSeqLogEntry ((decodeLEWire' ts ks <$> les) `using` parList rseq))
+          eRvr <- rseq (toSetRvr ((fromWire ts ks <$> vts) `using` parList rseq))
+          case eRvr of
+            Left !err -> return $! Left $! "Caught an invalid RVR in an AE: " ++ err
+            Right !vts' -> do
+              _ <- rseq eLes
+              case eLes of
+                Left !err -> return $! Left $ "Found a LogEntry with an invalid Command: " ++ err
+                Right !les' -> return $! Right $! AppendEntries t lid pli pt les' vts' $ ReceivedMsg dig bdy ts
+  {-# INLINE toWire #-}
+  {-# INLINE fromWire #-}
+
+data AppendEntriesResponse = AppendEntriesResponse
+  { _aerTerm       :: !Term
+  , _aerNodeId     :: !NodeID
+  , _aerSuccess    :: !Bool
+  , _aerConvinced  :: !Bool
+  , _aerIndex      :: !LogIndex
+  , _aerHash       :: !ByteString
+  , _aerWasVerified:: !Bool
+  , _aerProvenance :: !Provenance
+  }
+  deriving (Show, Generic, Eq)
+
+instance Ord AppendEntriesResponse where
+  -- This is here to get a set of AERs to order correctly
+  -- Node matters most (apples to apples)
+  -- Term supersedes index always
+  -- Index is really what we're after for how Set AER is used
+  -- Hash matters more than verified due to conflict resolution
+  -- Then verified, which is metadata really, because if everything up to that point is the same then the one that already ran through crypto is more valuable
+  -- After that it doesn't matter.
+  (AppendEntriesResponse t n s c i h v p) <= (AppendEntriesResponse t' n' s' c' i' h' v' p') =
+    (n,t,i,h,v,s,c,p) <= (n',t',i',h',v',s',c',p')
+
+data AERWire = AERWire (Term,NodeID,Bool,Bool,LogIndex,ByteString)
+  deriving (Show, Generic)
+instance Serialize AERWire
+
+instance WireFormat AppendEntriesResponse where
+  toWire nid pubKey privKey AppendEntriesResponse{..} = case _aerProvenance of
+    NewMsg -> let bdy = S.encode $ AERWire ( _aerTerm
+                                               , _aerNodeId
+                                               , _aerSuccess
+                                               , _aerConvinced
+                                               , _aerIndex
+                                               , _aerHash)
+                  sig = sign bdy privKey pubKey
+                  dig = Digest nid sig pubKey AER
+              in SignedRPC dig bdy
+    ReceivedMsg{..} -> SignedRPC _pDig _pOrig
+  fromWire !ts !ks s@(SignedRPC !dig !bdy) = case verifySignedRPC ks s of
+    Left !err -> Left $! err
+    Right () -> if _digType dig /= AER
+      then error $ "Invariant Failure: attempting to decode " ++ show (_digType dig) ++ " with AERWire instance"
+      else case S.decode bdy of
+        Left !err -> Left $! "Failure to decode AERWire: " ++ err
+        Right (AERWire !(t,nid,s',c,i,h)) -> Right $! AppendEntriesResponse t nid s' c i h True $ ReceivedMsg dig bdy ts
+  {-# INLINE toWire #-}
+  {-# INLINE fromWire #-}
+
+
+newtype AlotOfAERs = AlotOfAERs { _unAlot :: Map NodeID (Set AppendEntriesResponse)}
+  deriving (Show, Eq)
+
+instance Monoid AlotOfAERs where
+  mempty = AlotOfAERs Map.empty
+  mappend (AlotOfAERs m) (AlotOfAERs m') = AlotOfAERs $ Map.unionWith Set.union m m'
+
+data RequestVote = RequestVote
+  { _rvTerm        :: !Term
+  , _rvCandidateId :: !NodeID -- Sender ID Right? We don't forward RV's
+  , _lastLogIndex  :: !LogIndex
+  , _lastLogTerm   :: !Term
+  , _rvProvenance  :: !Provenance
+  }
+  deriving (Show, Eq, Generic)
+
+data RVWire = RVWire (Term,NodeID,LogIndex,Term)
+  deriving (Show, Generic)
+instance Serialize RVWire
+
+instance WireFormat RequestVote where
+  toWire nid pubKey privKey RequestVote{..} = case _rvProvenance of
+    NewMsg -> let bdy = S.encode $ RVWire ( _rvTerm
+                                          , _rvCandidateId
+                                          , _lastLogIndex
+                                          , _lastLogTerm)
+                  sig = sign bdy privKey pubKey
+                  dig = Digest nid sig pubKey RV
+              in SignedRPC dig bdy
+    ReceivedMsg{..} -> SignedRPC _pDig _pOrig
+  fromWire !ts !ks s@(SignedRPC !dig !bdy) = case verifySignedRPC ks s of
+    Left !err -> Left $! err
+    Right () -> if _digType dig /= RV
+      then error $ "Invariant Failure: attempting to decode " ++ show (_digType dig) ++ " with RVWire instance"
+      else case S.decode bdy of
+        Left !err -> Left $! "Failure to decode RVWire: " ++ err
+        Right (RVWire !(t,cid,lli,llt)) -> Right $! RequestVote t cid lli llt $ ReceivedMsg dig bdy ts
+  {-# INLINE toWire #-}
+  {-# INLINE fromWire #-}
+
+data RequestVoteResponse = RequestVoteResponse
+  { _rvrTerm        :: !Term
+  , _rvrCurLogIndex :: !LogIndex
+  , _rvrNodeId      :: !NodeID
+  , _voteGranted    :: !Bool
+  , _rvrCandidateId :: !NodeID
+  , _rvrProvenance  :: !Provenance
+  }
+  deriving (Show, Eq, Ord, Generic)
+
+data RVRWire = RVRWire (Term,LogIndex,NodeID,Bool,NodeID)
+  deriving (Show, Generic)
+instance Serialize RVRWire
+
+instance WireFormat RequestVoteResponse where
+  toWire nid pubKey privKey RequestVoteResponse{..} = case _rvrProvenance of
+    NewMsg -> let bdy = S.encode $ RVRWire (_rvrTerm,_rvrCurLogIndex,_rvrNodeId,_voteGranted,_rvrCandidateId)
+                  sig = sign bdy privKey pubKey
+                  dig = Digest nid sig pubKey RVR
+              in SignedRPC dig bdy
+    ReceivedMsg{..} -> SignedRPC _pDig _pOrig
+  fromWire !ts !ks s@(SignedRPC !dig !bdy) = case verifySignedRPC ks s of
+    Left !err -> Left $! err
+    Right () -> if _digType dig /= RVR
+      then error $ "Invariant Failure: attempting to decode " ++ show (_digType dig) ++ " with RVRWire instance"
+      else case S.decode bdy of
+        Left !err -> Left $! "Failure to decode RVRWire: " ++ err
+        Right (RVRWire !(t,li,nid,granted,cid)) -> Right $! RequestVoteResponse t li nid granted cid $ ReceivedMsg dig bdy ts
+  {-# INLINE toWire #-}
+  {-# INLINE fromWire #-}
+
+-- TODO: check if `toSetRvr eRvrs = Set.fromList <$> sequence eRvrs` is fusable?
+toSetRvr :: [Either String RequestVoteResponse] -> Either String (Set RequestVoteResponse)
+toSetRvr eRvrs = go eRvrs Set.empty
+  where
+    go [] s = Right $! s
+    go (Right rvr:rvrs) s = go rvrs (Set.insert rvr s)
+    go (Left err:_) _ = Left $! err
+{-# INLINE toSetRvr #-}
+
+-- the expected behavior here is tricky. For a set of votes, we are actually okay if some are invalid so long as there's a quorum
+-- however while we're still in alpha I think these failures represent a bug. Hence, they should be raised asap.
+decodeRVRWire :: Maybe ReceivedAt -> KeySet -> [SignedRPC] -> Either String (Set RequestVoteResponse)
+decodeRVRWire ts ks votes' = go votes' Set.empty
+  where
+    go [] s = Right $! s
+    go (v:vs) s = case fromWire ts ks v of
+      Left err -> Left $! err
+      Right rvr' -> go vs (Set.insert rvr' s)
+{-# INLINE decodeRVRWire #-}
+
+data Revolution = Revolution
+  { _revClientId   :: !NodeID
+  , _revLeaderId   :: !NodeID
+  , _revRequestId  :: !RequestId
+  , _revProvenance :: !Provenance
+  }
+  deriving (Show, Eq, Generic)
+
+data REVWire = REVWire (NodeID,NodeID,RequestId)
+  deriving (Show, Generic)
+instance Serialize REVWire
+
+instance WireFormat Revolution where
+  toWire nid pubKey privKey Revolution{..} = case _revProvenance of
+    NewMsg -> let bdy = S.encode $ REVWire (_revClientId,_revLeaderId,_revRequestId)
+                  sig = sign bdy privKey pubKey
+                  dig = Digest nid sig pubKey REV
+              in SignedRPC dig bdy
+    ReceivedMsg{..} -> SignedRPC _pDig _pOrig
+  fromWire !ts !ks s@(SignedRPC !dig !bdy) = case verifySignedRPC ks s of
+    Left !err -> Left $! err
+    Right () -> if _digType dig /= REV
+      then error $ "Invariant Failure: attempting to decode " ++ show (_digType dig) ++ " with REVWire instance"
+      else case S.decode bdy of
+        Left !err -> Left $! "Failure to decode REVWire: " ++ err
+        Right (REVWire !(cid,lid,rid)) -> Right $! Revolution cid lid rid $ ReceivedMsg dig bdy ts
+  {-# INLINE toWire #-}
+  {-# INLINE fromWire #-}
+
+
+data RPC = AE'   AppendEntries
+         | AER'  AppendEntriesResponse
+         | RV'   RequestVote
+         | RVR'  RequestVoteResponse
+         | CMD'  Command
+         | CMDB' CommandBatch
+         | CMDR' CommandResponse
+         | REV'  Revolution
+  deriving (Show, Generic)
+
+signedRPCtoRPC :: Maybe ReceivedAt -> KeySet -> SignedRPC -> Either String RPC
+signedRPCtoRPC ts ks s@(SignedRPC (Digest _ _ _ AE)   _) = (\rpc -> rpc `seq` AE'   rpc) <$> fromWire ts ks s
+signedRPCtoRPC ts ks s@(SignedRPC (Digest _ _ _ AER)  _) = (\rpc -> rpc `seq` AER'  rpc) <$> fromWire ts ks s
+signedRPCtoRPC ts ks s@(SignedRPC (Digest _ _ _ RV)   _) = (\rpc -> rpc `seq` RV'   rpc) <$> fromWire ts ks s
+signedRPCtoRPC ts ks s@(SignedRPC (Digest _ _ _ RVR)  _) = (\rpc -> rpc `seq` RVR'  rpc) <$> fromWire ts ks s
+signedRPCtoRPC ts ks s@(SignedRPC (Digest _ _ _ CMD)  _) = (\rpc -> rpc `seq` CMD'  rpc) <$> fromWire ts ks s
+signedRPCtoRPC ts ks s@(SignedRPC (Digest _ _ _ CMDR) _) = (\rpc -> rpc `seq` CMDR' rpc) <$> fromWire ts ks s
+signedRPCtoRPC ts ks s@(SignedRPC (Digest _ _ _ CMDB) _) = (\rpc -> rpc `seq` CMDB' rpc) <$> fromWire ts ks s
+signedRPCtoRPC ts ks s@(SignedRPC (Digest _ _ _ REV)  _) = (\rpc -> rpc `seq` REV'  rpc) <$> fromWire ts ks s
+{-# INLINE signedRPCtoRPC #-}
+
+rpcToSignedRPC :: NodeID -> PublicKey -> PrivateKey -> RPC -> SignedRPC
+rpcToSignedRPC nid pubKey privKey (AE' v) = toWire nid pubKey privKey v
+rpcToSignedRPC nid pubKey privKey (AER' v) = toWire nid pubKey privKey v
+rpcToSignedRPC nid pubKey privKey (RV' v) = toWire nid pubKey privKey v
+rpcToSignedRPC nid pubKey privKey (RVR' v) = toWire nid pubKey privKey v
+rpcToSignedRPC nid pubKey privKey (CMD' v) = toWire nid pubKey privKey v
+rpcToSignedRPC nid pubKey privKey (CMDR' v) = toWire nid pubKey privKey v
+rpcToSignedRPC nid pubKey privKey (CMDB' v) = toWire nid pubKey privKey v
+rpcToSignedRPC nid pubKey privKey (REV' v) = toWire nid pubKey privKey v
+{-# INLINE rpcToSignedRPC #-}
+
+data Event = ERPC RPC
+           | AERs AlotOfAERs
+           | ElectionTimeout String
+           | HeartbeatTimeout String
+  deriving (Show)
+
+data Role = Follower
+          | Candidate
+          | Leader
+  deriving (Show, Generic, Eq)
+
+data Metric
+  -- Consensus metrics:
+  = MetricTerm Term
+  | MetricCommitIndex LogIndex
+  | MetricCommitPeriod Double          -- For computing throughput
+  | MetricCurrentLeader (Maybe NodeID)
+  | MetricHash ByteString
+  -- Node metrics:
+  | MetricNodeId NodeID
+  | MetricRole Role
+  | MetricAppliedIndex LogIndex
+  | MetricApplyLatency Double
+  -- Cluster metrics:
+  | MetricClusterSize Int
+  | MetricQuorumSize Int
+  | MetricAvailableSize Int
+
+-- | A structure containing all the implementation details for running
+-- the raft protocol.
+-- Types:
+-- nt -- "node type", ie identifier (host/port, topic, subject)
+-- et -- "entry type", serialized format for submissions into Raft
+-- rt -- "return type", serialized format for "results" or "responses"
+-- mt -- "message type", serialized format for sending over wire
+data RaftSpec m = RaftSpec
+  {
+    -- ^ Function to get a log entry from persistent storage.
+    _readLogEntry     :: LogIndex -> m (Maybe CommandEntry) -- Simple [unused]
+
+    -- ^ Function to write a log entry to persistent storage.
+  , _writeLogEntry    :: LogIndex -> (Term,CommandEntry) -> m () -- Simple [unused]
+
+    -- ^ Function to get the term number from persistent storage.
+  , _readTermNumber   :: m Term -- Simple [unused]
+
+    -- ^ Function to write the term number to persistent storage.
+  , _writeTermNumber  :: Term -> m () -- Simple,Util(updateTerm[write only])
+
+    -- ^ Function to read the node voted for from persistent storage.
+  , _readVotedFor     :: m (Maybe NodeID) -- Simple [unused]
+
+    -- ^ Function to write the node voted for to persistent storage.
+  , _writeVotedFor    :: Maybe NodeID -> m () -- Simple,Role [write only]
+
+    -- ^ Function to apply a log entry to the state machine.
+  , _applyLogEntry    :: CommandEntry -> m CommandResult -- Simple,Handler
+
+    -- ^ Function to send a message to a node.
+  , _sendMessage      :: NodeID -> ByteString -> m () -- Simple,Sender
+
+    -- ^ Send more than one message at once
+  , _sendMessages     :: [(NodeID,ByteString)] -> m () -- Simple,Sender
+
+    -- ^ Function to get the next message.
+  , _getMessage       :: m (ReceivedAt, SignedRPC) -- Simple,Util(messageReceiver)
+
+    -- ^ Function to get the next N SignedRPCs not of type CMD, CMDB, or AER
+  , _getMessages   :: Int -> m [(ReceivedAt, SignedRPC)] -- Simple,Util(messageReceiver)
+
+    -- ^ Function to get the next N SignedRPCs of type CMD or CMDB
+  , _getNewCommands   :: Int -> m [(ReceivedAt, SignedRPC)] -- Simple,Util(messageReceiver)
+
+    -- ^ Function to get the next N SignedRPCs of type AER
+  , _getNewEvidence   :: Int -> m [(ReceivedAt, SignedRPC)] -- Simple,Util(messageReceiver)
+
+    -- ^ Function to log a debug message (no newline).
+  , _debugPrint       :: NodeID -> String -> m () -- Simple,Util(debug)
+
+  , _publishMetric    :: Metric -> m ()
+
+  , _getTimestamp     :: m UTCTime
+
+  , _random           :: forall a . Random a => (a, a) -> m a -- Simple,Util(randomRIO[timer])
+
+  , _enqueue          :: Event -> m () -- Simple,Util(enqueueEvent)
+
+  , _enqueueMultiple  :: [Event] -> m ()
+
+  , _enqueueLater     :: Int -> Event -> m ThreadId -- Simple,Util(enqueueEventLater[timer])
+
+  , _killEnqueued     :: ThreadId -> m () -- Simple,Timer
+
+  , _dequeue          :: m Event -- Simple,Util(dequeueEvent)
+  }
+makeLenses (''RaftSpec)
+
+data RaftState = RaftState
+  { _role             :: Role -- Handler,Role,Util(debug)
+  , _term             :: Term -- Handler,Role,Sender,Util(updateTerm)
+  , _votedFor         :: Maybe NodeID -- Handler,Role
+  , _lazyVote         :: Maybe (Term, NodeID, LogIndex) -- Handler
+  , _currentLeader    :: Maybe NodeID -- Client,Handler,Role
+  , _ignoreLeader     :: Bool -- Handler
+  , _logEntries       :: Log LogEntry -- Handler,Role,Sender
+  , _commitIndex      :: LogIndex -- Handler
+  , _lastApplied      :: LogIndex -- Handler
+  , _commitProof      :: Map NodeID AppendEntriesResponse -- Handler
+  , _timerThread      :: Maybe ThreadId -- Timer
+  , _timeSinceLastAER :: Int -- microseconds
+  , _replayMap        :: Map (NodeID, Signature) (Maybe CommandResult) -- Handler
+  , _cYesVotes        :: Set RequestVoteResponse -- Handler,Role,Sender
+  , _cPotentialVotes  :: Set NodeID -- Hander,Role,Sender
+  , _lNextIndex       :: Map NodeID LogIndex -- Handler,Role,Sender
+  , _lMatchIndex      :: Map NodeID LogIndex -- Role (never read?)
+  , _lConvinced       :: Set NodeID -- Handler,Role,Sender
+  , _lLastBatchUpdate :: (UTCTime, Maybe ByteString)
+  -- used for metrics
+  , _lastCommitTime   :: Maybe UTCTime
+
+  -- used by clients
+  , _pendingRequests  :: Map RequestId Command -- Client
+  , _currentRequestId :: RequestId -- Client
+  , _numTimeouts      :: Int -- Client
+  }
+makeLenses ''RaftState
+
+initialRaftState :: RaftState
+initialRaftState = RaftState
+  Follower   -- role
+  startTerm  -- term
+  Nothing    -- votedFor
+  Nothing    -- lazyVote
+  Nothing    -- currentLeader
+  False      -- ignoreLeader
+  mempty     -- log
+  startIndex -- commitIndex
+  startIndex -- lastApplied
+  Map.empty  -- commitProof
+  Nothing    -- timerThread
+  0          -- timeSinceLastAER
+  Map.empty  -- replayMap
+  Set.empty  -- cYesVotes
+  Set.empty  -- cPotentialVotes
+  Map.empty  -- lNextIndex
+  Map.empty  -- lMatchIndex
+  Set.empty  -- lConvinced
+  (minBound, Nothing)   -- lLastBatchUpdate (when we start up, we want batching to fire immediately)
+  Nothing    -- lastCommitTime
+  Map.empty  -- pendingRequests
+  0          -- nextRequestId
+  0          -- numTimeouts
+
+
+type Raft m = RWST (RaftEnv m) () RaftState m
+
+data RaftEnv m = RaftEnv
+  { _cfg         :: Config
+  , _clusterSize :: Int
+  , _quorumSize  :: Int  -- Handler,Role
+  , _rs          :: RaftSpec (Raft m)
+  }
+makeLenses ''RaftEnv
