+++ src/Juno/Runtime/Role.hs
@@ -5,6 +5,7 @@ module Juno.Runtime.Role
 
 import Juno.Runtime.Timer
 import Juno.Runtime.Types
+import Juno.Runtime.Protocol.Types
 import Juno.Util.Util
 
 -- THREAD: unknown/unused. updates state
