+++ src/Juno/Runtime/Sender.hs
@@ -27,6 +27,7 @@ import Data.Serialize
 
 import Juno.Util.Util
 import Juno.Runtime.Types
+import Juno.Runtime.Protocol.Types
 import Juno.Runtime.Timer (resetLastBatchUpdate)
 import Juno.Runtime.Log
